fix(script): rename file to adopt kebab-case
